pip-requirements.txt
python-telegram-bot==12.1.0
pillow==6.0.0
requests==2.21.0
xmltodict==0.12.0
beautifulsoup4==4.7.1
lxml==4.4.1
Django==2.2.7
